feat: Add evaluation harness (Milestone 4)
- Add dytopo-analyze crate for trace analysis and metrics
- Add `analyze` CLI command with text/json/csv output and SVG plots
- Add `benchmark` CLI command for dynamic vs baseline comparison
- Add baseline topologies: fully-connected, star, chain, ring
- Generate markdown benchmark reports
- SVG plots: score evolution, timing, density over rounds
- Fix u128 -> u64 for JSON serialization compatibility

+++ crates/core/src/lib.rs
@@ -43,7 +43,7 @@ pub enum TraceEvent {
         round: usize,
         goal: String,
         agent_count: usize,
-        ts_unix_ms: u128,
+        ts_unix_ms: u64,
     },
     AgentIO {
         round: usize,
@@ -65,6 +65,6 @@ pub enum TraceEvent {
     },
     RoundEnd {
         round: usize,
-        ts_unix_ms: u128,
+        ts_unix_ms: u64,
     },
 }
