+++ crates/orchestrator/src/lib.rs
@@ -29,11 +29,11 @@ pub struct RunArtifacts {
     pub messages: Vec<Message>,
 }
 
-fn now_ms() -> u128 {
+fn now_ms() -> u64 {
     SystemTime::now()
         .duration_since(UNIX_EPOCH)
         .unwrap_or_default()
-        .as_millis()
+        .as_millis() as u64
 }
 
 pub fn run_stub(
